Added header with implementation of the generator
Created a new header file with STB style implementation macro.
This header contains the implementation of the code generator, and it allows using the generator as a library if the user desires to make their own programs with it rather than using the command line tool issued by this repo.