Modified generator2.c to be a command line tool
The generator2.c file now includes the generator.h header file and makes use of the defined generator functionality to offer a command line program.

This program allows the user to generate custom counter macros with whatever configuration they choose.

Multiple counters can be generated in a single call to the generator program.