Added experimental alternative system
Created an experimental alternative system that works based on the idea of creating a set of digits and concatenating them rather than having a large #elif chain for every single possible counter value.

Each digit has a set of #elifs for the possible values of 0 to 9, leading to a far smaller overhead during preprocessing. This system also greatly reduces the number of required header files for a counter to work, potentially lowering it to a single header file (2 header files at most).