Added new source files for new version
Yes, I should have made some commits for the history to be logically tracked, but I coded all of this while on a laptop with no internet connection and away from my usual setup, so it is what it is.
TODO : Clean up and write a proper README with usage instructions. Also, remove or clean up the deprecated stuff.